# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Event Finder API - discover charity galas, silent auctions, and fundraising events
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get all admin settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
  /admin/email-recipients:
    get:
      operationId: getEmailRecipients
      tags: [admin]
      summary: List email recipients
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailRecipient"
    post:
      operationId: createEmailRecipient
      tags: [admin]
      summary: Add an email recipient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRecipientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailRecipient"
  /admin/email-recipients/{id}:
    patch:
      operationId: updateEmailRecipient
      tags: [admin]
      summary: Toggle active/inactive or update recipient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRecipientUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailRecipient"
    delete:
      operationId: deleteEmailRecipient
      tags: [admin]
      summary: Remove an email recipient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/domain-blacklist:
    get:
      operationId: getDomainBlacklist
      tags: [admin]
      summary: List blacklisted domains
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlacklistedDomain"
    post:
      operationId: addDomainBlacklist
      tags: [admin]
      summary: Add a domain to the blacklist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlacklistedDomainInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlacklistedDomain"
  /admin/domain-blacklist/{id}:
    delete:
      operationId: deleteDomainBlacklist
      tags: [admin]
      summary: Remove a domain from the blacklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Admin Users ──────────────────────────────────────────────────────────────
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create a user and send activation email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserCreated"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/users/{id}/resend-invite:
    post:
      operationId: resendAdminUserInvite
      tags: [admin]
      summary: Regenerate an activation link for a pending user (also attempts email)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteLinkResponse"
  /admin/users/{id}/reset-link:
    post:
      operationId: createAdminUserResetLink
      tags: [admin]
      summary: Generate a password reset link for an activated user (share out-of-band)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetLinkResponse"
...
components:
  schemas:
...
    AdminSettings:
      type: object
      required: [searchKeywords, avoidKeywords, overrideKeywords, maxPagesPerSite, timeoutPerPage, imageReadingEnabled, scoringWeights, tiers, minScoreForEmail]
      properties:
        searchKeywords:       { type: array, items: { type: string } }
        avoidKeywords:        { type: array, items: { type: string } }
        overrideKeywords:     { type: array, items: { type: string } }
        maxPagesPerSite:      { type: integer }
        timeoutPerPage:       { type: integer }
        imageReadingEnabled:  { type: boolean }
        scoringWeights:
          $ref: "#/components/schemas/ScoringWeights"
        tiers:
          type: array
          items:
            $ref: "#/components/schemas/ScoringTier"
        minScoreForEmail:     { type: integer }
        scheduleEnabled:      { type: boolean }
        scheduleDays:         { type: array, items: { type: string } }
        scheduleTime:         { type: string }
        scheduleTimezone:     { type: string }
        scheduleUrlListId:    { type: integer, nullable: true }
        primaryGroup1Label:   { type: string }
        primaryGroup3Label:   { type: string }
        defaultMailingState:  { type: string }
        geographicStates:     { type: array, items: { type: string } }
        maxConcurrentSites:   { type: integer }
        perDomainDelayMs:     { type: integer }
        maxRetries:           { type: integer }
    AdminSettingsUpdate:
      type: object
      properties:
        searchKeywords:       { type: array, items: { type: string } }
        avoidKeywords:        { type: array, items: { type: string } }
        overrideKeywords:     { type: array, items: { type: string } }
        maxPagesPerSite:      { type: integer }
        timeoutPerPage:       { type: integer }
        imageReadingEnabled:  { type: boolean }
        scoringWeights:
          $ref: "#/components/schemas/ScoringWeights"
        tiers:
          type: array
          items:
            $ref: "#/components/schemas/ScoringTier"
        minScoreForEmail:     { type: integer }
        scheduleEnabled:      { type: boolean }
        scheduleDays:         { type: array, items: { type: string } }
        scheduleTime:         { type: string }
        scheduleTimezone:     { type: string }
        scheduleUrlListId:    { type: integer, nullable: true }
        primaryGroup1Label:   { type: string }
        primaryGroup3Label:   { type: string }
        defaultMailingState:  { type: string }
        geographicStates:     { type: array, items: { type: string } }
        maxConcurrentSites:   { type: integer, minimum: 1, maximum: 50 }
        perDomainDelayMs:     { type: integer, minimum: 0, maximum: 60000 }
        maxRetries:           { type: integer, minimum: 0, maximum: 10 }
    ScoringTier:
      type: object
      required: [id, name, description, minScore]
      properties:
        id:          { type: string }
        name:        { type: string }
        description: { type: string }
        minScore:    { type: integer }
    ScoringWeights:
      type: object
      required: [hasAuction, auctionTypeBonus, highTicketPrice, midTicketPrice, lowTicketPrice, formalityBonus, timingPrime, timingGood, timingPenalty, beneficiaryBonus]
      properties:
        hasAuction:       { type: integer }
        auctionTypeBonus: { type: integer }
        highTicketPrice:  { type: integer }
        midTicketPrice:   { type: integer }
        lowTicketPrice:   { type: integer }
        formalityBonus:   { type: integer }
        timingPrime:      { type: integer }
        timingGood:       { type: integer }
        timingPenalty:    { type: integer }
        beneficiaryBonus: { type: integer }
    EmailRecipient:
      type: object
      required: [id, email, active, createdAt]
      properties:
        id:        { type: integer }
        email:     { type: string }
        active:    { type: boolean }
        createdAt: { type: string }
    EmailRecipientInput:
      type: object
      required: [email]
      properties:
        email: { type: string }
    EmailRecipientUpdate:
      type: object
      properties:
        email:  { type: string }
        active: { type: boolean }
    BlacklistedDomain:
      type: object
      required: [id, domain, createdAt]
      properties:
        id:        { type: integer }
        domain:    { type: string }
        createdAt: { type: string }
    BlacklistedDomainInput:
      type: object
      required: [domain]
      properties:
        domain: { type: string }

    # ─── Auth ─────────────────────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:          { type: integer }
        name:        { type: string }
        email:       { type: string }
        role:        { type: string }
        activatedAt: { type: ["string", "null"] }
        createdAt:   { type: string }
    AdminUserInput:
      type: object
      required: [name, email]
      properties:
        name:  { type: string }
        email: { type: string }
        role:  { type: string }
    AdminUserUpdate:
      type: object
      properties:
        name:  { type: string }
        email: { type: string }
        role:  { type: string }
    AdminUserCreated:
      type: object
      required: [id, name, email, role, createdAt, activationUrl, emailSent]
      properties:
        id:            { type: integer }
        name:          { type: string }
        email:         { type: string }
        role:          { type: string }
        activatedAt:   { type: ["string", "null"] }
        createdAt:     { type: string }
        activationUrl: { type: string }
        emailSent:     { type: boolean }
    InviteLinkResponse:
      type: object
      required: [message, activationUrl, emailSent]
      properties:
        message:       { type: string }
        activationUrl: { type: string }
        emailSent:     { type: boolean }
    ResetLinkResponse:
      type: object
      required: [resetUrl, expiresAt]
      properties:
        resetUrl:  { type: string }
        expiresAt: { type: string }
...